Improve handling of 64-bit immediates on x86.
Before this change, operations like Omulimm were allowing 64-bit
immediates while the instruction set does not. As a consequence, the x86
printer was forced to store the immediates in memory and to generate
memory loads.

Now, only signed 32-bit immediates (which are sign-extended during
instruction decoding) are allowed. Larger immediates are delgated to the
register allocator and memory loads are no longer generated.